+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # CHANGELOG
 
 <!-- version list -->
+## v0.0.1c0 (2026-05-14)
+- Add Class MySolensoStationCount,
+- Add Class MySolensoPowerByDay,
+- Add Class MySolensoCountByDayOfYeay,
+- Correction auth - function get_auth_headers_hoymiles
+
+
 ## v0.0.1b0 (2026-05-14)
 - Add Run PyTest to build,
-- Add function station_find,
+- Add class MySolensoStationData,
 - Update documentation.
 
 ## v0.0.1a0 (2026-05-14)